move commodity dictionary to own file, parse independently of Cyclus input file
