+++ src/main.py
@@ -1,5 +1,6 @@
 from argparse import ArgumentParser
 import parse_input as pi
+import commodity_dictionary as cd
 
 
 def make_parser():
@@ -18,7 +19,7 @@ def main(args=None):
     input_file = 'source_3_sink_1.xml'
 
     if (input_file) is not None:
-        commodity_dictionary = pi.parse_input(input_file)
+        commodity_dictionary = cd.build_commod_dictionary()
     else:
         print('No input file given!')
